chore(trading): move wait for graphql to utils
