+++ apps/trading/e2e/tests/chart_depth/test_chart_depth.py
@@ -1,7 +1,7 @@
 import pytest
 from playwright.sync_api import expect, Page
 
-
+#REMOVE
 @pytest.mark.usefixtures("page", "continuous_market", "risk_accepted")
 def test_see_market_depth_chart(continuous_market, page: Page):
     page.goto(f"/#/markets/{continuous_market}")
