+++ apps/trading/e2e/tests/deal_ticket/test_stop_order_oco.py
@@ -1,4 +1,3 @@
-from math import exp
 import pytest
 from collections import namedtuple
 from playwright.sync_api import Page, expect
@@ -6,7 +5,7 @@ from vega_sim.service import VegaService
 from actions.vega import submit_order
 from conftest import init_vega, page
 from fixtures.market import setup_continuous_market
-from actions.utils import wait_for_toast_confirmation
+from actions.utils import wait_for_toast_confirmation, wait_for_graphql_response
 
 # Defined namedtuples
 WalletConfig = namedtuple("WalletConfig", ["name", "passphrase"])
@@ -53,33 +52,6 @@ oco = "oco"
 trigger_price_oco = "triggerPrice-oco"
 order_size_oco = "order-size-oco"
 order_limit_price_oco = "order-price-oco"
-
-
-def wait_for_graphql_response(page, query_name, timeout=5000):
-    response_data = {}
-
-    def handle_response(route, request):
-        if "graphql" in request.url:
-            response = request.response()
-            if response is not None:
-                json_response = response.json()
-                if json_response and "data" in json_response:
-                    data = json_response["data"]
-                    if query_name in data:
-                        response_data["data"] = data
-                        route.continue_()
-                        return
-        route.continue_()
-
-    # Register the route handler
-    page.route("**", handle_response)
-
-    # Wait for the response data to be populated
-    page.wait_for_timeout(timeout)
-
-    # Unregister the route handler
-    page.unroute("**", handle_response)
-
 
 def create_position(vega: VegaService, market_id):
     submit_order(vega, "Key 1", market_id, "SIDE_SELL", 100, 110)
