+++ apps/trading/e2e/tests/trade_history/test_trade_history.py
@@ -1,9 +1,12 @@
 import pytest
 import re
+import logging
 from playwright.sync_api import expect
 from actions.vega import submit_order
 from conftest import init_vega
-import logging
+from actions.utils import wait_for_graphql_response
+from playwright.sync_api import Page
+from vega_sim.null_service import VegaService
 
 logger = logging.getLogger()
 
@@ -15,7 +18,7 @@ def vega():
 
 
 # Could be turned into a helper function in the future.
-def verify_data_grid(page, data_test_id, expected_pattern):
+def verify_data_grid(page: Page, data_test_id, expected_pattern):
     page.get_by_test_id(data_test_id).click()
     # Required so that we can get liquidation price
     if data_test_id == "Positions":
@@ -45,35 +48,8 @@ def verify_data_grid(page, data_test_id, expected_pattern):
                 raise AssertionError(f"Pattern does not match: {expected} != {actual}")
 
 
-# Required so that we can get liquidation price - Could also become a helper
-def wait_for_graphql_response(page, query_name, timeout=5000):
-    response_data = {}
-
-    def handle_response(route, request):
-        if "graphql" in request.url:
-            response = request.response()
-            if response is not None:
-                json_response = response.json()
-                if json_response and "data" in json_response:
-                    data = json_response["data"]
-                    if query_name in data:
-                        response_data["data"] = data
-                        route.continue_()
-                        return
-        route.continue_()
-
-    # Register the route handler
-    page.route("**", handle_response)
-
-    # Wait for the response data to be populated
-    page.wait_for_timeout(timeout)
-
-    # Unregister the route handler
-    page.unroute("**", handle_response)
-
-
 @pytest.mark.usefixtures("page", "continuous_market", "auth", "risk_accepted")
-def test_limit_order_new_trade_top_of_list(continuous_market, vega, page):
+def test_limit_order_new_trade_top_of_list(continuous_market, vega: VegaService, page: Page):
     submit_order(vega, "Key 1", continuous_market, "SIDE_BUY", 1, 110)
     page.goto(f"/#/markets/{continuous_market}")
 
@@ -97,7 +73,7 @@ def test_limit_order_new_trade_top_of_list(continuous_market, vega, page):
 
 
 @pytest.mark.usefixtures("page", "continuous_market", "auth", "risk_accepted")
-def test_price_copied_to_deal_ticket(continuous_market, page):
+def test_price_copied_to_deal_ticket(continuous_market, page: Page):
     page.goto(f"/#/markets/{continuous_market}")
     page.get_by_test_id("Trades").click()
     wait_for_graphql_response(page, "Trades")
